Revert "oscam-whitelist: Fix missing sha256sum checksum warning"
The checksum was ommitted on purpose, the repo only contains one file, and that changes regularly. For that reason, BB_STRICT_CHECKSUM has been disabled to so the missing checksum becomes a warning, instead of stopping the build (which an incorrect checksum would do too).

Afaik there is no way around this issue.

This reverts commit 2fa6affbcfbe771718a2f98c3d76d436022738d6.